requirements.txt
alignment==1.0.10
asciitree==0.3.3
beautifulsoup4==4.8.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
Cython==0.29.28
fasteners==0.17.3
idna==2.8
imagecodecs==2022.2.22
keras==2.9.0
libtiff==0.4.2
numcodecs==0.9.1
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
qpsolvers==1.9.1
quadprog==0.1.11
requests==2.22.0
scipy==1.8.1
soupsieve==2.3.2.post1
tifffile==2022.4.8
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.25.11
zarr==2.11.3
